Link against my DWARF parsing library
*** This requires an external repository.  Run make to get it. ***

Since this lives in a separate repository, the makefile automatically
checks for the repository and checks that a known-good version is
checked out.  This is roughly equivalent to git submodules, but it's
integrated into the build process and uses commit author dates instead
of commit hashes since they're stable against history revision and
self-describing.  The makefile will also automatically build the DWARF
library if it is out of date.