Support "streaming" JSON dictionaries and lists
These write themselves out to an ostream on-the-fly.  Currently, this
is used for the top-level output dictionary so you can see each report
as it's generated.

The API is less-than-ideal since it has to support a mix of streaming
and non-streaming objects, but I don't particularly want to port all
of mscan to streaming output.