Use streaming output within the asharing analysis reports
The asharing reports can be really big, especially when you mess
something up.  This dumps them to stdout as they are generated.  In
particular, the sharing analysis and summary information are now
generated up-front so they can be output immediately.  This doesn't
take very long and then the rest of the output can be slowly generated
from the computed sharing results.