More robust and simpler JSON streaming logic
Now everything is streaming, but objects can be buffered if they
aren't the current object in the streaming order.  The output stream
acts as a token that gets passed between objects in a depth-first
traversal order.  It can be held at a collection object if that
collection is not yet complete.

I wish I had tail recursion.